Revamp: Pundit authZ, public/privileged separation, migration integrity, demo seed, green test harness
- Stage 0: Remove fatal/dead code (SoundCloud LoadError, Facebook OAuth secret leak + before_filter); quarantine dead controllers
- Stage 1: Wire Pundit (include Pundit::Authorization), retire RoleBasedAccess global gate; public index/show open, mutations authorized
- Stage 2: Gate all backstage actions; close unauth-write holes (artists/paintings/feature_banners/tickets)
- Stage 3: Fix 57 corrupted migrations + bigint FK bug; 97/97 migrate from scratch; regenerate schema
- Stage 4: Seed demo content (mall/store/3 products/event); add missing slug columns (malls/stores/products) + mall views
- Stage 5: Fix test harness (green: 9 runs/48 assertions/0 failures); DRY role helpers into RoleBasedAccess; env-driven seed password; remove dead CarrierWave uploader
- Board reviews: BOARD_REVIEW_STAGE_0_1 / _2 / _3 / _4 / _5 / _6_FINAL (6/6 approve, 0 blockers)

+++ db/migrate/20251102163600_create_ads.rb
@@ -5,7 +5,7 @@ class CreateAds < ActiveRecord::Migration[7.2]
       t.text :image
       t.text :image_data
       t.string :url
-      t.integer :ad_spot_id, null: false
+      t.bigint :ad_spot_id, null: false
       t.date :start_date
       t.date :end_date
       t.boolean :active, default: true, null: false
